# backend/services/gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: Platform API Gateway
  version: 1.0.0
  description: |
    Unified API Gateway for the multi-tenant cloud deployment platform.
    All frontend traffic is routed through this gateway which handles:
    - Authentication (JWT and service account tokens)
    - Authorization (organization and project scope)
    - Rate limiting
    - Request correlation
    - Distributed tracing
servers:
  - url: http://localhost:8080/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /auth/verify-email:
    post:
      tags: [Auth]
      operationId: verifyEmail
      summary: Verify email address
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
      responses:
        "200":
          description: Email verified
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/password-reset:
    post:
      tags: [Auth]
      operationId: requestPasswordReset
      summary: Request password reset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "200":
          description: Reset email sent
  /auth/password-reset/confirm:
    post:
      tags: [Auth]
      operationId: confirmPasswordReset
      summary: Confirm password reset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmPasswordResetRequest"
      responses:
        "200":
          description: Password reset
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/mfa/setup:
    post:
      tags: [Auth]
      operationId: setupMFA
      summary: Start MFA setup
      responses:
        "200":
          description: MFA setup initiated
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/mfa/enable:
    post:
      tags: [Auth]
      operationId: enableMFA
      summary: Enable MFA with TOTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MFAEnableRequest"
      responses:
        "200":
          description: MFA enabled
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/mfa/disable:
    post:
      tags: [Auth]
      operationId: disableMFA
      summary: Disable MFA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MFADisableRequest"
      responses:
        "200":
          description: MFA disabled
        "400":
          $ref: "#/components/responses/BadRequest"

  # ============================================================================
  # Tenant Service Routes
  # ============================================================================
...
  /organizations/{orgId}/audit-logs:
    parameters:
      - $ref: "#/components/parameters/OrgId"
    get:
      tags: [Audit]
      operationId: listAuditLogs
      summary: List audit logs
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: domain
          in: query
          schema:
            type: string
          description: Filter by event domain (e.g., auth, tenant, project)
        - name: eventType
          in: query
          schema:
            type: string
          description: Filter by event type
        - name: startTime
          in: query
          schema:
            type: string
            format: date-time
          description: Filter events after this time
        - name: endTime
          in: query
          schema:
            type: string
            format: date-time
          description: Filter events before this time
      responses:
        "200":
          description: Audit log list
  /organizations/{orgId}/audit-logs/{eventId}:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: Event ID
    get:
      tags: [Audit]
      operationId: getAuditLog
      summary: Get audit log entry
      responses:
        "200":
          description: Audit log entry
        "404":
          $ref: "#/components/responses/NotFound"

  # ============================================================================
  # Service Account Routes
  # ============================================================================
  /organizations/{orgId}/service-accounts:
    parameters:
      - $ref: "#/components/parameters/OrgId"
    post:
      tags: [Service Accounts]
      operationId: createServiceAccount
      summary: Create service account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceAccountRequest"
      responses:
        "201":
          description: Service account created
    get:
      tags: [Service Accounts]
      operationId: listServiceAccounts
      summary: List service accounts
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Service account list
  /organizations/{orgId}/service-accounts/{accountId}:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - name: accountId
        in: path
        required: true
        schema:
          type: string
        description: Service account ID
    get:
      tags: [Service Accounts]
      operationId: getServiceAccount
      summary: Get service account
      responses:
        "200":
          description: Service account details
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Service Accounts]
      operationId: deleteServiceAccount
      summary: Delete service account
      responses:
        "204":
          description: Service account deleted

  # ============================================================================
  # API Token Routes
  # ============================================================================
  /organizations/{orgId}/api-tokens:
    parameters:
      - $ref: "#/components/parameters/OrgId"
    post:
      tags: [API Tokens]
      operationId: createAPIToken
      summary: Create API token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAPITokenRequest"
      responses:
        "201":
          description: API token created
    get:
      tags: [API Tokens]
      operationId: listAPITokens
      summary: List API tokens
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: API token list
  /organizations/{orgId}/api-tokens/{tokenId}:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - name: tokenId
        in: path
        required: true
        schema:
          type: string
        description: API token ID
    delete:
      tags: [API Tokens]
      operationId: revokeAPIToken
      summary: Revoke API token
      responses:
        "204":
          description: API token revoked
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by the auth service
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Organization ID
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of items to return
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Pagination cursor from previous response
  schemas:
...
    VerifyEmailRequest:
      type: object
      properties:
        token:
          type: string
      required: [token]
    PasswordResetRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    ConfirmPasswordResetRequest:
      type: object
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 8
      required: [token, newPassword]
    MFAEnableRequest:
      type: object
      properties:
        totpCode:
          type: string
      required: [totpCode]
    MFADisableRequest:
      type: object
      properties:
        totpCode:
          type: string
      required: [totpCode]
...
    CreateServiceAccountRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        scopes:
          type: array
          items:
            type: string
      required: [name]
    CreateAPITokenRequest:
      type: object
      properties:
        name:
          type: string
        expiresIn:
          type: string
          description: Duration like "30d" or "1y"
        scopes:
          type: array
          items:
            type: string
      required: [name]
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        requestId:
          type: string
      required: [error]
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...